speed up --help
